feat: Add BusinessErrorsInterceptor
Add a new interceptor, BusinessErrorsInterceptor, to handle business errors in the application. This interceptor catches errors and throws appropriate HTTP exceptions based on the error type. If the error type is BusinessError.NOT_FOUND, it throws a 404 Not Found exception. If the error type is BusinessError.PRECONDITION_FAILED, it throws a 412 Precondition Failed exception. If the error type is BusinessError.BAD_REQUEST, it throws a 400 Bad Request exception. For any other error type, it rethrows the error as is.